% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnPredict.R
\name{nn_predict}
\alias{nn_predict}
\title{nnpredict - Perform a single forward run of the neural network model.}
\usage{
nn_predict(model_name, input_vec, echo = FALSE)
}
\arguments{
\item{model_name}{Name of the model that defines the neural network in iQuest format.}

\item{input_vec}{Numeric vector with the same length as the number of input nodes.}

\item{echo}{Toggle printing of debug statements within neural network C code.}
}
\value{
Numeric vector of neural network model outputs corresponding to the inputs provided.
}
\description{
nnpredict - Perform a single forward run of the neural network model.
}
\examples{
model_filename   <- file.path(system.file("extdata/WI_StreamTemp.enn", package="nnlib") )
static_filename  <- file.path(system.file("extdata/static_data.csv", package="nnlib") )
dynamic_filename <- file.path(system.file("extdata/dynamic_data.csv", package="nnlib") )
static_data  <- read.csv( static_filename, header=TRUE )
dynamic_data <- read.csv( dynamic_filename, header=TRUE )

# select a row from each of the static and dynamic data files
input_vector <- as.numeric( c( static_data[ 37, 3:10], dynamic_data[ 3, c(2,3,4,5,6,7,10,16,18)] ) )

# now run the forward neural network model with a single days' worth of data
stream_temp  <- nn_predict( model_name=model_filename, input_vec=input_vector )

}

