% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnPredictTS.R
\name{nn_predict_ts}
\alias{nn_predict_ts}
\title{nn_predict_ts - Perform a series of neural network predictions for a set of dynamic model inputs.}
\usage{
nn_predict_ts(model_name, num_static_fields, num_dyn_fields, input_static_vec,
  input_dyn_vec, echo = FALSE, field_order = NA)
}
\arguments{
\item{model_name}{Name of the model that defines the neural network in iQuest format.}

\item{num_static_fields}{Number of static neural network inputs to be processed.}

\item{num_dyn_fields}{Number of dynamic neural network inputs to be processed.}

\item{input_static_vec}{Vector of static neural network model intputs.}

\item{input_dyn_vec}{Vector of dynamic neural network model inputs; length will be equal to the number
of dynamic fields times the number of repetitions (i.e. days) worth of data included in the vector.}

\item{echo}{Print debug information within the neural network C code if TRUE.}

\item{field_order}{Dynamic inputs can be rearranged according to the field_order vector.}
}
\value{
Matrix of output values; columns represent values of the output nodes of the ANN. Number of rows
        will equal the length of the input dynamic vector (input_dyn_vec).
}
\description{
nn_predict_ts - Perform a series of neural network predictions for a set of dynamic model inputs.
}
\examples{
model_filename   <- file.path(system.file("extdata/WI_StreamTemp.enn", package="nnlib") )
static_filename  <- file.path(system.file("extdata/static_data.csv", package="nnlib") )
dynamic_filename <- file.path(system.file("extdata/dynamic_data.csv", package="nnlib") )
static_data  <- read.csv( static_filename, header=TRUE )
dynamic_data <- read.csv( dynamic_filename, header=TRUE )

# select a row from each of the static and dynamic data files
static_input_vector  <- as.numeric( static_data[ 37, 3:10] )
dynamic_input_matrix <- dynamic_data[ 3:7, c( 2, 3, 4, 5, 6, 7, 8, 13, 15) ]
dynamic_input_vector <- as.vector( t( dynamic_input_matrix ) )

# now run the forward neural network model with multiple days' worth of data
stream_temps  <- nn_predict_ts( model_name=model_filename,
                                num_static_fields=8,
                                num_dyn_fields=9,
                                input_static_vec=static_input_vector,
                                input_dyn_vec=dynamic_input_vector )

}

